+++ Code/Mantid/Framework/Kernel/src/VectorHelper.cpp
@@ -102,8 +102,8 @@ void rebin(const std::vector<double>& xold, const std::vector<double>& yold, con
   if (size_xnew != (ynew.size() + 1) || size_xnew != (enew.size() + 1))
     throw std::runtime_error("rebin: y and error vectors should be of same size & 1 shorter than x");
 
-  int size_yold = yold.size();
-  int size_ynew = ynew.size();
+  size_t size_yold = yold.size();
+  size_t size_ynew = ynew.size();
 
   if (!addition)
   {
